Moved dispatching logic into ResourceWrapper class
Resources/handlers are now more light-weight. Handlers are found by inspection
of the resource object for properties that have a '_rhino_meta' attribute which
contains the request handler info.

All resources added to a mapper are automatically wrapped in a ResourceWrapper
instance, which finds the handlers and contains all the dispatching logic that
was previously in dispatch_request(). As an optimization, Mapper instances are
not wrapped.

Special cases:
- A resource can also be a standalone handler function
- A resource can be a class. It will be instantiated before handling a
  request
- A resource can implement __call__, which will be called with request
  and ctx (if requested) as arguments to implement custom initialization.
  In this case __call__ must not return a value.
- If a resource has no handlers, it is called with request and ctx (if
  requested) and is expected to return a Response instance or None
  (raises 404).

Caveat: We now longer can make any assumptions about the relative
ordering of handlers inside a resource, e.g. for resolving ties during
content-negotiation.