Resources can now override how URLs are built
If a resource (the target to `Mapper.add`) implements `build_url`,
this function is used to translate the parameters passed to
`Request.url_for` to the parameters expected by the route's URL
template. `build_url` should return a path component by calling
the url building function passed in as first arguments with
positional parameters required by the template.